ia funcional executas varias acoes

+++ app/src/main/java/ao/easy/vvia/interfaces/ResourceHandler.java
@@ -3,5 +3,5 @@ package ao.easy.vvia.interfaces;
 import android.content.Context;
 
 public interface ResourceHandler {
-    void execute(Context context, int value);
+    void execute(Context context, double value);
 }
